pyproject.toml
[project]
name = "dynamic-civ"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "pygame-ce>=2.5.7",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
